update: update suspense in `/blogs` components
