--- src/app/blogs/blog-grid-skeleton.tsx
@@ -1,9 +1,0 @@
-"use client";
-
-import * as React from "react";
-
-const BlogGridSkeleton: React.FC = () => {
-  return <div></div>;
-};
-
-export default BlogGridSkeleton;
