+++ src/app/blogs/layout.tsx
@@ -1,20 +1,13 @@
-import * as React from "react";
-import { Metadata } from "next";
-
-export const metadata: Metadata = {
-  title: "Blogs",
-};
+import React, { type JSX } from "react";
 
 type Props = {
   children: React.ReactNode;
 };
 
-const Blog: React.FC<Props> = async ({ children }) => {
+export default function BlogsLayout({ children }: Props): JSX.Element {
   return (
     <div className="mt-8 w-screen max-w-full sm:mt-12 md:mt-20 lg:mt-28">
-      {children}
+      <React.Suspense fallback={null}>{children}</React.Suspense>
     </div>
   );
-};
-
-export default Blog;
+}
